refactor: 代码重构与优化
主要更改:
- 修复中文注释乱码问题
- 优化头文件包含结构,减少循环依赖
- 为所有函数添加详细的文档注释
- 修复类型不匹配问题(uint8->uint32)
- 改进菜单系统的滚动显示逻辑
- 增强相机显示模式,添加赛道线条可视化
- 优化图像处理流程,添加完整的处理步骤
- 移除servo_motor_duty的重复赋值
- 统一代码风格和注释格式

+++ code/pid.h
@@ -68,8 +68,8 @@ extern float gyro_gain_scale;  // 角速度环死区增益缩放
 extern float steer_kp;           // 转向P系数（基于图像偏差）
 extern float steer_kd;           // 转向D系数（基于陀螺仪gz）
 extern float steer_output_limit; // 转向输出限幅
-extern uint8 steer_sample_start; // 图像采样起始行
-extern uint8 steer_sample_end;   // 图像采样结束行
+extern uint32 steer_sample_start; // 图像采样起始行
+extern uint32 steer_sample_end;   // 图像采样结束行
 
 // 输出平滑参数（导出到菜单）
 extern float output_filter_coeff; // 输出滤波系数
